update: clean up files and change naming conventions of the files
