+++ examples/offline_inference/moss_tts_local/end2end.py
@@ -2,11 +2,11 @@
 Offline TTS inference with MOSS-TTS-Local on vllm-omni.
 
 Usage:
-  cd /workspace/vllm-omni/examples/offline_inference/moss_tts
+  cd /workspace/vllm-omni/examples/offline_inference/moss_tts_local
   MOSS_AUDIO_TOKENIZER_PATH=/workspace/vllm-omni/weights/moss-audio-tokenizer \
   python end2end.py \
     --model /workspace/vllm-omni/weights/moss-tts-local \
-    --stage-configs-path ../../../vllm_omni/model_executor/stage_configs/moss_tts.yaml \
+    --stage-configs-path ../../../vllm_omni/model_executor/stage_configs/moss_tts_local.yaml \
     --text "The weather is so nice today." \
     --output-dir ./output_audio
 """
@@ -51,19 +51,6 @@ def build_tts_prompt(
     ambient_sound=None,
     language=None,
 ) -> str:
-    """
-    Build the full TTS generation prompt.
-
-    Replicates what MossTTSDelayProcessor.__call__(mode='generation') does
-    for a single-turn TTS request (no reference audio):
-
-      1. Format content with <user_inst>...</user_inst> template
-      2. Wrap with chat template  →  <|im_start|>user ... <|im_end|><|im_start|>assistant
-      3. Append <|audio_start|>   →  signals start of audio generation
-
-    The AR stage model generates <|audio_assistant_gen_slot|> tokens after
-    <|audio_start|>; each one triggers the local transformer to emit RVQ codes.
-    """
     from transformers import AutoTokenizer
 
     tok = AutoTokenizer.from_pretrained(
@@ -73,25 +60,22 @@ def build_tts_prompt(
 
     content = (
         _USER_INST_TEMPLATE
-        .replace("{reference}",    "None")
-        .replace("{instruction}",  str(instruction))
-        .replace("{tokens}",       str(tokens))
-        .replace("{quality}",      str(quality))
-        .replace("{sound_event}",  str(sound_event))
-        .replace("{ambient_sound}",str(ambient_sound))
-        .replace("{language}",     str(language))
-        .replace("{text}",         str(text))
+        .replace("{reference}", "None")
+        .replace("{instruction}", str(instruction))
+        .replace("{tokens}", str(tokens))
+        .replace("{quality}", str(quality))
+        .replace("{sound_event}", str(sound_event))
+        .replace("{ambient_sound}", str(ambient_sound))
+        .replace("{language}", str(language))
+        .replace("{text}", str(text))
     )
 
-    # Chat-template wrapping (adds <|im_start|>...<|im_end|> + generation prompt)
     prompt = tok.apply_chat_template(
         [{"role": "user", "content": content}],
         tokenize=False,
         add_generation_prompt=True,
     )
 
-    # Append <|audio_start|> — replicates the audio_start_position_tokens append
-    # in MossTTSDelayProcessor.__call__() for generation mode.
     prompt = prompt + "<|audio_start|>"
     return prompt
 
@@ -99,20 +83,17 @@ def build_tts_prompt(
 def main():
     import argparse
     parser = argparse.ArgumentParser()
-    parser.add_argument("--model",
-                        default="/workspace/vllm-omni/weights/moss-tts-local")
-    parser.add_argument("--stage-configs-path",
-                        default="../../../vllm_omni/model_executor/stage_configs/moss_tts.yaml")
-    parser.add_argument("--text",        default="The weather is so nice today.")
-    parser.add_argument("--output-dir",  default="./output_audio")
+    parser.add_argument("--model", default="/workspace/vllm-omni/weights/moss-tts-local")
+    parser.add_argument("--stage-configs-path", default="../../../vllm_omni/model_executor/stage_configs/moss_tts_local.yaml")
+    parser.add_argument("--text", default="The weather is so nice today.")
+    parser.add_argument("--output-dir", default="./output_audio")
     parser.add_argument("--num-prompts", type=int, default=1)
     parser.add_argument("--init-sleep-seconds", type=int, default=20)
-    parser.add_argument("--batch-timeout",       type=int, default=5)
-    parser.add_argument("--init-timeout",        type=int, default=5000)
+    parser.add_argument("--batch-timeout", type=int, default=5)
+    parser.add_argument("--init-timeout", type=int, default=5000)
     parser.add_argument("--shm-threshold-bytes", type=int, default=65536)
     args = parser.parse_args()
 
-    # ── Codec path ───────────────────────────────────────────────────────────
     if not os.environ.get("MOSS_AUDIO_TOKENIZER_PATH"):
         sibling = os.path.join(
             os.path.dirname(os.path.abspath(args.model)),
@@ -126,21 +107,16 @@ def main():
 
     os.makedirs(args.output_dir, exist_ok=True)
 
-    # ── Prompt ───────────────────────────────────────────────────────────────
     prompt = build_tts_prompt(args.text, args.model)
     print(f"[Info] Prompt ({len(prompt)} chars):\n{prompt}\n")
 
-    # ── Resolve EOS / stop tokens from the model config ─────────────────────
-    # The AR stage stops when it samples <|audio_end|> (audio_end_token_id)
-    # which signals the model has finished generating audio frames.
-    # We also include the standard text EOS so generation never overshoots.
     from transformers import AutoConfig
     _cfg = AutoConfig.from_pretrained(
         os.path.abspath(args.model), trust_remote_code=True
     )
     _audio_end_id = getattr(_cfg, "audio_end_token_id", None)
-    _eos_id       = getattr(_cfg, "eos_token_id",       None)
-    # eos_token_id may be a list (Qwen3 has two EOS tokens)
+    _eos_id       = getattr(_cfg, "eos_token_id", None)
+
     if isinstance(_eos_id, list):
         _eos_ids = _eos_id
     elif _eos_id is not None:
@@ -153,27 +129,24 @@ def main():
     print(f"[Info] AR stop token IDs: {ar_stop_ids}  "
           f"(audio_end={_audio_end_id}, eos={_eos_ids})")
 
-    # ── Sampling params ──────────────────────────────────────────────────────
-    # AR stage (stage 0): each generated token produces 32 RVQ codes fed to
-    # stage 1. Stage 1 has max_model_len=18192, so the flat code sequence must
-    # stay within that limit: max_ar_tokens = floor(18192 / 32) = 568.
-    # Use 500 as a hard ceiling; EOS stopping should end generation much sooner.
+
     ar_params = SamplingParams(
-        temperature=0.6, top_p=0.95, top_k=50,
-        max_tokens=500, seed=SEED,
-        # NOTE: do NOT set repetition_penalty here.
-        # The gen_slot token (151656) must repeat every single decode step.
-        # A repetition penalty ≠ 1.0 suppresses it after the first occurrence,
-        # causing the backbone to emit random vocabulary tokens instead of
-        # gen_slot tokens, which produces garbled/random audio content.
+        temperature=0.6, 
+        top_p=0.95, 
+        top_k=50,
+        max_tokens=500, 
+        seed=SEED,
         stop_token_ids=ar_stop_ids if ar_stop_ids else None,
     )
     decoder_params = SamplingParams(
-        temperature=0.0, top_p=1.0, top_k=-1,
-        max_tokens=18192, seed=SEED, detokenize=False,
+        temperature=0.0, 
+        top_p=1.0, 
+        top_k=-1,
+        max_tokens=18192, 
+        seed=SEED, 
+        detokenize=False,
     )
 
-    # ── Engine ───────────────────────────────────────────────────────────────
     omni = Omni(
         model=args.model,
         stage_configs_path=args.stage_configs_path,
@@ -187,7 +160,6 @@ def main():
     print(f"[Info] Running {len(prompts)} prompt(s)...")
     omni_outputs = omni.generate(prompts, [ar_params, decoder_params])
 
-    # ── Collect outputs ──────────────────────────────────────────────────────
     for stage_outputs in omni_outputs:
         output = stage_outputs.request_output
         rid    = output.request_id
