--- vllm_omni/model_executor/stage_configs/moss_tts.yaml
@@ -1,116 +1,0 @@
-# Stage config for MOSS-TTS-Local — batch (non-streaming) mode.
-#
-# Two-stage pipeline
-# ──────────────────
-#  Stage 0  "ar_stage"   Global Qwen3-1.7B + Local Transformer → 32-layer RVQ codes
-#  Stage 1  "decoder"    CAT codec (1.6 B)                     → 24 kHz waveform
-#
-# Verified hardware target: single H20-96G (96 GB HBM).
-# For smaller GPUs, reduce gpu_memory_utilization accordingly.
-# For multi-GPU, set tensor_parallel_size and adjust devices.
-#
-# Environment variables
-# ─────────────────────
-#   MOSS_AUDIO_TOKENIZER_PATH   Path to the CAT codec checkpoint.
-#                               If unset, falls back to the main model directory.
-
-async_chunk: false
-
-stage_args:
-  # ════════════════════════════════════════════════════════════════════
-  #  Stage 0 — AR stage  (Qwen3 global backbone + local transformer)
-  # ════════════════════════════════════════════════════════════════════
-  - stage_id: 0
-    stage_type: llm
-    runtime:
-      process: true
-      devices: "0" # CUDA_VISIBLE_DEVICES for this process
-
-    engine_args:
-      dtype: bfloat16
-      max_num_seqs: 1
-      model_stage: ar_stage
-      model_arch: MossTTSForConditionalGeneration
-      worker_type: ar
-      scheduler_cls: vllm_omni.core.sched.omni_ar_scheduler.OmniARScheduler
-      tensor_parallel_size: 1 # scale up for multi-GPU (e.g. 2 for 2×GPU)
-
-      # Memory budget
-      # Qwen3-1.7B global backbone ≈ 3.4 GB @ bf16
-      # Local transformer (4 blocks, 1536 hidden) ≈ 0.5 GB @ bf16
-      # Embeddings + LM heads (33 channels) ≈ 0.3 GB @ bf16
-      # Total AR stage ≈ 4.2 GB
-      gpu_memory_utilization: 0.3
-
-      enforce_eager: true # local transformer runs eagerly (no CUDA graph)
-      trust_remote_code: true # required for MossTTSLocalTransformer import
-      # MUST stay false: each audio code stream is unique and not a stable
-      # prefix.  Caching would corrupt downstream waveforms across requests.
-      enable_prefix_caching: false
-      engine_output_type: latent
-      max_model_len: 4096
-      max_num_batched_tokens: 4096
-      async_scheduling: false
-
-    # `is_comprehension: true` on Stage 0 means it consumes a user prompt
-    # (vs Stage 1 which only consumes upstream stage output).
-    is_comprehension: true
-    final_output: true
-    final_output_type: text # text tokens surfaced for debugging; audio in Stage 1
-
-    default_sampling_params:
-      temperature: 0.6 # text channel temperature
-      top_p: 0.95
-      top_k: 50
-      max_tokens: 300
-      seed: 42
-      detokenize: true
-      # NOTE: no repetition_penalty — gen_slot (151656) repeats every step.
-      # Stop at <|audio_end|> (151653) so the model doesn't run to max_tokens.
-      # Also include <|im_end|> (151645) and <|endoftext|> (151643) as fallbacks.
-      stop_token_ids: [151653, 151645, 151643]
-
-  # ════════════════════════════════════════════════════════════════════
-  #  Stage 1 — Decoder  (CAT codec: RVQ codes → 24 kHz waveform)
-  # ════════════════════════════════════════════════════════════════════
-  - stage_id: 1
-    stage_type: llm
-    runtime:
-      process: true
-      devices: "0" # CAT codec on a separate GPU; set "0" for single-GPU
-
-    engine_args:
-      model_stage: decoder
-      max_num_seqs: 1
-      model_arch: MossTTSForConditionalGeneration
-      worker_type: generation
-      scheduler_cls: vllm_omni.core.sched.omni_generation_scheduler.OmniGenerationScheduler
-      tensor_parallel_size: 1
-
-      # Memory budget
-      # CAT codec (1.6 B parameters) ≈ 3.2 GB @ bf16
-      # Activation memory during decode ≈ 1 GB
-      # Total ≈ 4.2 GB
-      gpu_memory_utilization: 0.25
-
-      enforce_eager: true
-      trust_remote_code: true
-      enable_prefix_caching: false
-      engine_output_type: audio
-      max_model_len: 18192
-      max_num_batched_tokens: 18192
-      async_scheduling: false
-
-    engine_input_source: [0]
-    is_comprehension: false
-    custom_process_input_func: vllm_omni.model_executor.stage_input_processors.moss_tts.llm2decoder
-    final_output: true
-    final_output_type: audio
-
-    default_sampling_params:
-      temperature: 0.0 # decoder is deterministic
-      top_p: 1.0
-      top_k: -1
-      max_tokens: 18192
-      seed: 42
-      detokenize: false
