--- vllm_omni/model_executor/stage_configs/moss_tts_async.yaml
@@ -1,171 +1,0 @@
-# Stage config for MOSS-TTS-Local with async_chunk streaming.
-#
-# Scope
-# -----
-#   - offline / batch mode AND async-chunk streaming (this file = streaming)
-#   - direct TTS prompts
-#   - no reference-audio continuation
-#
-# Async-chunk overview
-# --------------------
-# Stage 0 (AR stage) emits one frame-major RVQ row [n_vq] per decode step.
-# llm2decoder_async_chunk() (set via custom_process_next_stage_input_func on
-# Stage 0) accumulates those rows and flushes codec_chunk_frames newly
-# generated frames to Stage 1 (CAT codec) via the shared-memory connector.
-
-# Stage 1 decodes each chunk to 24 kHz audio using the codec's streaming
-# KV-cache, without waiting for the full generation to finish.
-#
-# Local (no-delay) latency note
-# -----------------------------
-# Unlike the delay variant there is no ramp-up penalty — the first chunk
-# needs only `codec_chunk_frames` decode steps.  Keep the value small
-# (e.g. 3) to minimise initial audio latency.
-#
-# Hardware
-# --------
-# Same target as moss_tts.yaml (single H20-96G).  The CAT codec stage is
-# shared with the batch/sync mode and reuses the MOSS-Audio-Tokenizer
-# checkpoint resolved via MOSS_AUDIO_TOKENIZER_PATH.
-
-async_chunk: true
-
-stage_args:
-  # ════════════════════════════════════════════════════════════════════
-  #  Stage 0 — AR stage (Qwen3 global backbone + local transformer)
-  # ════════════════════════════════════════════════════════════════════
-  - stage_id: 0
-    stage_type: llm
-    runtime:
-      process: true
-      devices: "0"
-
-    engine_args:
-      dtype: bfloat16
-      max_num_seqs: 1
-      model_stage: ar_stage
-      model_arch: MossTTSForConditionalGeneration
-      worker_type: ar
-      scheduler_cls: vllm_omni.core.sched.omni_ar_scheduler.OmniARScheduler
-      tensor_parallel_size: 1
-      gpu_memory_utilization: 0.3
-      enforce_eager: true
-      trust_remote_code: true
-      enable_prefix_caching: false
-      engine_output_type: latent
-      max_model_len: 4096
-      max_num_batched_tokens: 4096
-      async_scheduling: true
-      custom_process_next_stage_input_func: vllm_omni.model_executor.stage_input_processors.moss_tts.llm2decoder_async_chunk
-
-    # Names a transport channel from this stage to Stage 1.  The right-hand
-    # side ("connector_of_shared_memory") refers to the connector defined
-    # under `runtime.connectors` at the bottom of this file.
-    output_connectors:
-      to_stage_1: connector_of_shared_memory
-
-    is_comprehension: true
-    final_output: true
-    final_output_type: text
-
-    default_sampling_params:
-      temperature: 0.6
-      top_p: 0.95
-      top_k: 50
-      max_tokens: 300
-      seed: 42
-      detokenize: true
-      # Stop at <|audio_end|> (151653) + standard EOS fallbacks.
-      stop_token_ids: [151653, 151645, 151643]
-
-  # ════════════════════════════════════════════════════════════════════
-  #  Stage 1 — Decoder (CAT codec: RVQ codes → 24 kHz waveform)
-  # ════════════════════════════════════════════════════════════════════
-  - stage_id: 1
-    stage_type: llm
-    runtime:
-      process: true
-      devices: "0"
-
-    engine_args:
-      model_stage: decoder
-      max_num_seqs: 1
-      model_arch: MossTTSForConditionalGeneration
-      worker_type: generation
-      scheduler_cls: vllm_omni.core.sched.omni_generation_scheduler.OmniGenerationScheduler
-      tensor_parallel_size: 1
-      gpu_memory_utilization: 0.25
-      enforce_eager: true
-      trust_remote_code: true
-      enable_prefix_caching: false
-      engine_output_type: audio
-      max_model_len: 18192
-      max_num_batched_tokens: 18192
-      # Stage 1 also schedules asynchronously so it can decode chunks as
-      # they arrive instead of blocking on a full input.
-      async_scheduling: true
-
-    # Stage 1 reads from Stage 0 — index `[0]` is "first stage in the pipeline".
-    # In streaming mode, the actual transport is the connector below; in batch
-    # mode (moss_tts.yaml) the runtime invokes `custom_process_input_func` on
-    # the stage at index 0 instead.
-    engine_input_source: [0]
-    # Counterpart to Stage 0's `output_connectors`.  The shared name
-    # ("connector_of_shared_memory") binds the two stages to the same
-    # connector instance defined under `runtime.connectors`.
-    input_connectors:
-      from_stage_0: connector_of_shared_memory
-    is_comprehension: false
-    final_output: true
-    final_output_type: audio
-
-    default_sampling_params:
-      temperature: 0.0
-      top_p: 1.0
-      top_k: -1
-      max_tokens: 18192
-      seed: 42
-      detokenize: false
-
-runtime:
-  enabled: true
-  defaults:
-    window_size: -1
-    max_inflight: 1
-
-  connectors:
-    # Inter-stage transport used by both `output_connectors` (Stage 0) and
-    # `input_connectors` (Stage 1).  The same dict instance is bound to both
-    # ends; payloads written here are read on the other side.
-    connector_of_shared_memory:
-      name: SharedMemoryConnector # transport implementation registered in vllm-omni
-      extra:
-        # Switch to /dev/shm-backed buffers once a payload exceeds this size,
-        # to avoid the cost of pickling large tensors through the IPC queue.
-        shm_threshold_bytes: 65536
-        # Engages the CAT codec's streaming KV-cache mode in Stage 1
-        # (consumed by MossTTSDecoderModel._enter_streaming).
-        codec_streaming: true
-        # Stage 1 polls the connector every connector_get_sleep_s seconds
-        # while waiting for a chunk.  Smaller = lower latency, more CPU.
-        connector_get_sleep_s: 0.01
-        # Max wait (ms) for the FIRST chunk of a request — long because
-        # Stage 0 prefill + first decode is slow on cold weights.
-        connector_get_max_wait_first_chunk: 3000
-        # Max wait (ms) for SUBSEQUENT chunks — tighter because steady-state
-        # decode is fast and a stall here means Stage 0 has crashed.
-        connector_get_max_wait: 300
-        # Local (no-delay) first chunk needs just codec_chunk_frames steps.
-        # This must match (or be the source of) the chunk size that
-        # `llm2decoder_async_chunk` uses to flush; the bridge function reads
-        # it via `connector.config.extra.codec_chunk_frames`.
-        codec_chunk_frames: 3
-        # MossTTSDecoderModel does not trim left context — the codec's
-        # internal KV-cache already carries the necessary context across
-        # chunks, so there is no overlap region to skip on this side.
-        codec_left_context_frames: 0
-
-  edges:
-    - from: 0
-      to: 1
-      window_size: -1
