+++ vllm_omni/model_executor/stage_configs/moss_tts_local_async.yaml
@@ -1,0 +1,104 @@
+# Stage config for MOSS-TTS-Local (Async Mode)
+
+async_chunk: true
+
+stage_args:
+  - stage_id: 0
+    stage_type: llm
+    runtime:
+      process: true
+      devices: "0"
+
+    engine_args:
+      dtype: bfloat16
+      max_num_seqs: 1
+      model_stage: ar_stage
+      model_arch: MossTTSForConditionalGeneration
+      worker_type: ar
+      scheduler_cls: vllm_omni.core.sched.omni_ar_scheduler.OmniARScheduler
+      tensor_parallel_size: 1
+      gpu_memory_utilization: 0.3
+      enforce_eager: true
+      trust_remote_code: true
+      enable_prefix_caching: false
+      engine_output_type: latent
+      max_model_len: 4096
+      max_num_batched_tokens: 4096
+      async_scheduling: true
+      custom_process_next_stage_input_func: vllm_omni.model_executor.stage_input_processors.moss_tts_local.llm2decoder_async_chunk
+
+    output_connectors:
+      to_stage_1: connector_of_shared_memory
+
+    is_comprehension: true
+    final_output: true
+    final_output_type: text
+
+    default_sampling_params:
+      temperature: 0.6
+      top_p: 0.95
+      top_k: 50
+      max_tokens: 300
+      seed: 42
+      detokenize: true
+      stop_token_ids: [151653, 151645, 151643]
+
+  - stage_id: 1
+    stage_type: llm
+    runtime:
+      process: true
+      devices: "0"
+
+    engine_args:
+      model_stage: decoder
+      max_num_seqs: 1
+      model_arch: MossTTSForConditionalGeneration
+      worker_type: generation
+      scheduler_cls: vllm_omni.core.sched.omni_generation_scheduler.OmniGenerationScheduler
+      tensor_parallel_size: 1
+      gpu_memory_utilization: 0.25
+      enforce_eager: true
+      trust_remote_code: true
+      enable_prefix_caching: false
+      engine_output_type: audio
+      max_model_len: 18192
+      max_num_batched_tokens: 18192
+      async_scheduling: true
+
+    engine_input_source: [0]
+    input_connectors:
+      from_stage_0: connector_of_shared_memory
+    is_comprehension: false
+    final_output: true
+    final_output_type: audio
+
+    default_sampling_params:
+      temperature: 0.0
+      top_p: 1.0
+      top_k: -1
+      max_tokens: 18192
+      seed: 42
+      detokenize: false
+
+runtime:
+  enabled: true
+  defaults:
+    window_size: -1
+    max_inflight: 1
+
+  connectors:
+    connector_of_shared_memory:
+      name: SharedMemoryConnector
+      extra:
+        shm_threshold_bytes: 65536
+        codec_streaming: true
+        connector_get_sleep_s: 0.01
+        connector_get_max_wait_first_chunk: 3000
+        connector_get_max_wait: 300
+        codec_chunk_frames: 3
+        codec_left_context_frames: 0
+
+  edges:
+    - from: 0
+      to: 1
+      window_size: -1
